+++ takoyaki_bin/src/commands/mod.rs
@@ -1,5 +1,7 @@
 mod init;
+mod plug;
 
 // Reexport
 pub use init::*;
+pub use plug::*;
