+++ takoyaki_bin/src/utils/metadata.rs
