+++ takoyaki_bin/src/utils/mod.rs
@@ -1,8 +1,10 @@
 mod command;
+mod metadata;
 mod log;
 mod helpers;
 
 pub use command::*;
 pub use log::*;
 pub use helpers::*;
+pub use metadata::*;
